requirements.txt
PyQt5==5.12.2
pycryptodome==3.8.2
cryptography==2.7